Refabricação das clases com code smell Data Class.
