+++ src/main/java/org/jfree/data/statistics/DefaultStatisticalCategoryDataset.java
@@ -21,7 +21,7 @@
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301,
  * USA.
  *
- * [Oracle and Java are registered trademarks of Oracle and/or its affiliates. 
+ * [Oracle and Java are registered trademarks of Oracle and/or its affiliates.
  * Other names may be trademarks of their respective owners.]
  *
  * --------------------------------------
@@ -58,6 +58,7 @@
 
 package org.jfree.data.statistics;
 
+import java.io.Serializable;
 import java.util.List;
 import java.util.Objects;
 import org.jfree.chart.internal.CloneUtils;
@@ -73,7 +74,7 @@ import org.jfree.data.general.DatasetChangeEvent;
  * A convenience class that provides a default implementation of the
  * {@link StatisticalCategoryDataset} interface.
  */
-public class DefaultStatisticalCategoryDataset<R extends Comparable<R>, 
+public class DefaultStatisticalCategoryDataset<R extends Comparable<R>,
         C extends Comparable<C>>  extends AbstractDataset
         implements StatisticalCategoryDataset<R, C>, RangeInfo, PublicCloneable {
 
@@ -353,9 +354,9 @@ public class DefaultStatisticalCategoryDataset<R extends Comparable<R>,
      * @param rowKey  the row key.
      * @param columnKey  the column key.
      */
-    public void add(double mean, double standardDeviation, R rowKey, 
-            C columnKey) {
-        add(Double.valueOf(mean), Double.valueOf(standardDeviation), rowKey, 
+    public void add(double mean, double standardDeviation, R rowKey,
+                    C columnKey) {
+        add(Double.valueOf(mean), Double.valueOf(standardDeviation), rowKey,
                 columnKey);
     }
 
@@ -744,4 +745,141 @@ public class DefaultStatisticalCategoryDataset<R extends Comparable<R>,
         clone.data = CloneUtils.clone(this.data);
         return clone;
     }
+
+    private class MeanAndStandardDeviation implements Serializable {
+
+        /** For serialization. */
+        private static final long serialVersionUID = 7413468697315721515L;
+
+        /** The mean. */
+        private Number mean;
+
+        /** The standard deviation. */
+        private Number standardDeviation;
+
+        /**
+         * Creates a new mean and standard deviation record.
+         *
+         * @param mean  the mean.
+         * @param standardDeviation  the standard deviation.
+         */
+        public MeanAndStandardDeviation(double mean, double standardDeviation) {
+            this(Double.valueOf(mean), Double.valueOf(standardDeviation));
+        }
+
+        /**
+         * Creates a new mean and standard deviation record.
+         *
+         * @param mean  the mean ({@code null} permitted).
+         * @param standardDeviation  the standard deviation ({@code null}
+         *                           permitted.
+         */
+        public MeanAndStandardDeviation(Number mean, Number standardDeviation) {
+            this.mean = mean;
+            this.standardDeviation = standardDeviation;
+        }
+
+        /**
+         * Returns the mean.
+         *
+         * @return The mean.
+         */
+        public Number getMean() {
+            return this.mean;
+        }
+
+        /**
+         * Returns the mean as a double primitive.  If the underlying mean is
+         * {@code null}, this method will return {@code Double.NaN}.
+         *
+         * @return The mean.
+         *
+         * @see #getMean()
+         *
+         * @since 1.0.7
+         */
+        public double getMeanValue() {
+            double result = Double.NaN;
+            if (this.mean != null) {
+                result = this.mean.doubleValue();
+            }
+            return result;
+        }
+
+        /**
+         * Returns the standard deviation.
+         *
+         * @return The standard deviation.
+         */
+        public Number getStandardDeviation() {
+            return this.standardDeviation;
+        }
+
+        /**
+         * Returns the standard deviation as a double primitive.  If the underlying
+         * standard deviation is {@code null}, this method will return
+         * {@code Double.NaN}.
+         *
+         * @return The standard deviation.
+         *
+         * @since 1.0.7
+         */
+        public double getStandardDeviationValue() {
+            double result = Double.NaN;
+            if (this.standardDeviation != null) {
+                result = this.standardDeviation.doubleValue();
+            }
+            return result;
+        }
+
+        /**
+         * Tests this instance for equality with an arbitrary object.
+         *
+         * @param obj  the object ({@code null} permitted).
+         *
+         * @return A boolean.
+         */
+        @Override
+        public boolean equals(Object obj) {
+            if (obj == this) {
+                return true;
+            }
+            if (!(obj instanceof DefaultStatisticalCategoryDataset.MeanAndStandardDeviation)) {
+                return false;
+            }
+            MeanAndStandardDeviation that = (MeanAndStandardDeviation) obj;
+            if (!Objects.equals(this.mean, that.mean)) {
+                return false;
+            }
+            if (!Objects.equals(this.standardDeviation, that.standardDeviation)
+            ) {
+                return false;
+            }
+            return true;
+        }
+
+        @Override
+        public int hashCode()
+        {
+            int hash = 3;
+            hash = 79 * hash + Objects.hashCode( this.mean );
+            hash = 79 * hash + Objects.hashCode( this.standardDeviation );
+            return hash;
+        }
+
+        /**
+         * Returns a string representing this instance.
+         *
+         * @return A string.
+         *
+         * @since 1.0.7
+         */
+        @Override
+        public String toString() {
+            return "[" + this.mean + ", " + this.standardDeviation + "]";
+        }
+
+    }
+
+
 }
